refactor: split robinhood watchlist read/write tools (#68)
