+++ src/open_stocks_mcp/tools/watchlists/__init__.py
@@ -1,0 +1,1 @@
+"""Watchlist tool modules split by read/write concerns."""
